+++ django_test_app/server/apps/jwt_auth/views.py
@@ -43,6 +43,16 @@ class ObtainAccessAndRefreshSyncController(
             ),
         }
 
+    @override
+    def login(self, parsed_body: ObtainTokensPayload) -> ObtainTokensResponse:
+        """This is needed only for test purpose."""
+        response = super().login(parsed_body)
+        # Testing:
+        assert (  # noqa: S101, PT018
+            self.request.user.is_authenticated and self.request.user.is_active
+        )
+        return response
+
 
 class ObtainAccessAndRefreshAsyncController(
     ObtainTokensAsyncController[
@@ -71,6 +81,21 @@ class ObtainAccessAndRefreshAsyncController(
                 token_type='refresh',  # noqa: S106
             ),
         }
+
+    @override
+    async def login(
+        self,
+        parsed_body: ObtainTokensPayload,
+    ) -> ObtainTokensResponse:
+        """This is needed only for test purpose."""
+        response = await super().login(parsed_body)
+        # Testing:
+        assert (  # noqa: S101, PT018
+            self.request.user.is_authenticated and self.request.user.is_active
+        )
+        auser = await self.request.auser()
+        assert auser.is_authenticated and auser.is_active  # noqa: S101, PT018
+        return response
 
 
 @final